LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/input-targets-root-while-render-blocked-expected.txt


FAIL Input when rendering suppressed targets root assert_equals: Events must target the transition root while render blocked expected Element node <html><head><title>View transitions: ensure input targets... but got Element node <div id="clicktarget"></div>
